fix table schema

+++ src/database/migrations/2024_06_09_154547_alter_user_table_add_field_status.php
@@ -15,7 +15,7 @@ class AlterUserTableAddFieldStatus extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('status')->after('role_id')->default(Constant::STATUS_ACTIVE);
+            $table->tinyInteger('status')->after('role_id')->default(Constant::STATUS_ACTIVE);
         });
     }
