+++ src/database/migrations/2024_06_17_085311_alter_table_purchase_order_item_add_field_unit.php
@@ -15,7 +15,7 @@ class AlterTablePurchaseOrderItemAddFieldUnit extends Migration
     public function up()
     {
         Schema::table('purchase_order_items', function (Blueprint $table) {
-            $table->unsignedInteger('unit')->nullable(false)->default(Constant::UNIT_PCS)->after('price');
+            $table->smallInteger('unit')->nullable(false)->default(Constant::UNIT_PCS)->after('price');
         });
     }
